refactor(parser): split lint.rs into lint/ module directory
Modularize the 4991-line lint.rs into focused submodules for
maintainability. No logic, behavior, or API changes -- pure
restructuring with identical test results (241 tests pass).

New structure:
  lint/
    mod.rs              -- public types, helpers, public API, config/suppression
    rules/
      mod.rs            -- submodule declarations
      metadata.rs       -- shared metadata checks (title, id, status, dates, etc.)
      detection.rs      -- detection rule checks, logsource, modifier compatibility
      correlation.rs    -- correlation rule checks, timespan validation
      filter.rs         -- filter rule checks
      shared.rs         -- unknown key detection, known-key constants

Roadmap item #14 (Large File Modularization), step 1/13.
