+++ crates/rsigma-parser/src/lint/rules/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod correlation;
+pub(crate) mod detection;
+pub(crate) mod filter;
+pub(crate) mod metadata;
+pub(crate) mod shared;
